% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{oxford_languages}
\title{Oxford Dictionaries API Constants}
\format{An object of class \code{character} of length 7.}
\usage{
oxford_languages
}
\description{
Some variables are defined in the constants folder to assist 
  in maintaining error checks or other utilities.
}
\keyword{datasets}
