% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxford_filters.R
\name{get_oxford_filters}
\alias{get_oxford_filters}
\title{Get the Languages Supported by the Oxford API}
\usage{
get_oxford_filters(endpoint = NULL, app_id = getOption("oxford_api_app_id"),
  app_key = getOption("oxford_api_app_key"),
  url_base = getOption("oxford_api_url_base"))
}
\arguments{
\item{endpoint}{\code{character(1)} or \code{NULL}. If given, the filters
for that endpoint are retrieved.}

\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{url_base}{\code{character(1)}. The url stem for the API.}
}
\value{
Returns a data frame describing the filters available in the API.

\itemize{
 \item \code{endpoint } The feature for which the filter is available.
 \item \code{filter } The name of the filter.
}
}
\description{
Returns a data frame of the filters supported by the 
  Oxford Dictionaries API>
}
\section{Functional Requirements}{

\enumerate{
 \item Return a data frame describing the filters available in the API.
 \item Correctly returns only the filters for the indicated endpoint
   when \code{endpoint} is not \code{NULL}.
 \item Cast an error if the API returns an error.
 \item Cast an error if \code{endpoint} is not \code{character(1)} or
   \code{NULL}.
 \item Cast an error if \code{app_id} is not \code{character(1)}.
 \item Cast an error if \code{app_id} is \code{NULL}.
 \item Cast an error if \code{app_key} is not \code{character(1)}.
 \item Cast an error if \code{app_key} is \code{NULL}.
 \item Cast an error if \code{url_base} is not \code{character(1)}.
}
}

\examples{
\dontrun{
set_application_access(app_id = [your_app_id],
                       app_key = [your_app_key])
get_oxford_languages()
}

}
