% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxford_inflections.R
\name{get_oxford_inflections}
\alias{get_oxford_inflections}
\title{Check if a Word Exists in the Dictionary and Retrieve its Root Form}
\usage{
get_oxford_inflections(word_id, ..., app_id = getOption("oxford_api_app_id"),
  app_key = getOption("oxford_api_app_key"),
  language = getOption("oxford_api_language"),
  url_base = getOption("oxford_api_url_base"))
}
\arguments{
\item{word_id}{\code{character(1)} The word to search for in the dictionary.}

\item{...}{Named arguments of character vectors giving additional filters
to apply to the search.}

\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{language}{\code{character(1)}. The language in which to search.}

\item{url_base}{\code{character(1)}. The url stem for the API.}
}
\value{
Returns a data frame
\itemize{
 \item \code{search_term_id } The search term's word ID
 \item \code{search_term } The text of the search term.
 \item \code{language } The language for which the search was conducted
 \item \code{lexical_category } The lexical category of the inflection.
 \item \code{grammatical_feature } The grammatical feature of the inflection.
 \item \code{grammatical_feature_type } The type of grammatical feature.
 \item \code{inflection_id } The word ID of the word for which the search
   term is an inflection.
 \item \code{inflection_text } The text of the word for which the search 
   term is an inflection.
}
}
\description{
"Use this to check if a word exists in the dictionary, or what
'root' form it links to (e.g., swimming > swim). The response tells you
the possible lemmas for a given inflected word. This can then be combined
with other endpoints to retrieve more information." (Source: Oxford 
Dictionaries API Documentation)
}
\section{Functional Requirements}{

\enumerate{
 \item Return a data frame with the results of the API call.
 \item Cast an error if \code{word_id} is not \code{character(1)}.
 \item Cast an error if any argument to \code{...} is not named.
 \item Cast an error if any argument to \code{...} is not a character 
   vector.
 \item Cast an error if the API returns an error.
 \item Cast an error if \code{app_id} is not \code{character(1)}.
 \item Cast an error if \code{app_id} is \code{NULL}.
 \item Cast an error if \code{app_key} is not \code{character(1)}.
 \item Cast an error if \code{app_key} is \code{NULL}.
 \item Cast an error if \code{language} is not one of 
   \code{oxford_languages}.
 \item Cast an error if \code{url_base} is not \code{character(1)}.
}
}

\examples{
\dontrun{
get_oxford_inflections(word_id = "swimming")

get_oxford_inflections(word_id = "change",
                       grammaticalFeatures = c("singular", "past"),
                       lexicalCategory = "noun")
}

}
