% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxford_lexical_categories.R
\name{get_oxford_lexical_categories}
\alias{get_oxford_lexical_categories}
\title{Get Domains for a Chosen Language}
\usage{
get_oxford_lexical_categories(app_id = getOption("oxford_api_app_id"),
  app_key = getOption("oxford_api_app_key"),
  language = getOption("oxford_api_language"),
  url_base = getOption("oxford_api_url_base"))
}
\arguments{
\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{language}{\code{character(1)}. The language for which to look up 
domains.}

\item{url_base}{\code{character(1)}. The url stem for the API.}
}
\value{
Returns a character vector of lexical categories in the 
  chosen language.
}
\description{
Returns the lexical categories available for the 
  selected language.
}
\section{Functional Requirements}{

\enumerate{
 \item Return a character vector of domains in the chosen language.
 \item Cast an error if the API returns an error.
 \item Cast an error if \code{app_id} is not \code{character(1)}.
 \item Cast an error if \code{app_id} is \code{NULL}.
 \item Cast an error if \code{app_key} is not \code{character(1)}.
 \item Cast an error if \code{app_key} is \code{NULL}.
 \item Cast an error if \code{language} is not \code{character(1)}.
 \item Cast an error if \code{url_base} is not \code{character(1)}.
}
}

\examples{
\dontrun{
set_application_access(app_id = [your_app_id],
                       app_key = [your_app_key])
get_oxford_lexical_categories()
}

}
