% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxford_thesaurus.R
\name{get_oxford_thesaurus}
\alias{get_oxford_thesaurus}
\title{Get Thesaurus Entries from the Oxford Dictionaries API}
\usage{
get_oxford_thesaurus(word_id, synonym = TRUE, antonym = FALSE,
  app_id = getOption("oxford_api_app_id"),
  app_key = getOption("oxford_api_app_key"),
  language = getOption("oxford_api_language"),
  url_base = getOption("oxford_api_url_base"))
}
\arguments{
\item{word_id}{\code{character(1)}. The word to search in the dictionary.}

\item{synonym}{\code{logical(1)}. If \code{TRUE}, return synonyms for 
\code{word_id}.}

\item{antonym}{\code{logical(1)}. If \code{FALSE}, return antonyms for
\code{word_id}.}

\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{language}{\code{character(1)}. The language in which to search.}

\item{url_base}{\code{character(1)}. The url stem for the API.}
}
\value{
Returns a data frame:
\itemize{
 \item \code{search_term } The term searched in the dictionary.
 \item \code{thesaurus_entry } A term returned from the dictionary.
 \item \code{type } Character value indicating if \code{thesaurus_entry}
   is a synonym or an antonym.
}
}
\description{
Retrieve words that are similar or opposite a given word.
}
\section{Functional Requirements}{

\enumerate{
 \item Return a data frame with the results of the API call.
 \item Cast an error if \code{word_id} is not \code{character(1)}.
 \item Cast an error if \code{synonym} is not \code{logical(1)}.
 \item Cast an error if \code{antonym} is not \code{logical(1)}.
 \item Return an empty data frame with a warning if both \code{synonym}
   and \code{antonym} are \code{FALSE}.
 \item Cast an error if the API returns an error.
 \item Cast an error if \code{app_id} is not \code{character(1)}.
 \item Cast an error if \code{app_id} is \code{NULL}.
 \item Cast an error if \code{app_key} is not \code{character(1)}.
 \item Cast an error if \code{app_key} is \code{NULL}.
 \item Cast an error if \code{language} is not one of 
   \code{oxford_languages}.
 \item Cast an error if \code{url_base} is not \code{character(1)}.
}
}

