% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oxford_wordlist.R
\name{get_oxford_wordlist}
\alias{get_oxford_wordlist}
\title{Get a List of Words}
\usage{
get_oxford_wordlist(..., advanced_filter = list(),
  app_id = getOption("oxford_api_app_id"),
  app_key = getOption("oxford_api_app_key"),
  language = getOption("oxford_api_language"),
  url_base = getOption("oxford_api_url_base"))
}
\arguments{
\item{...}{named arguments of character vectors to pass as filters to 
the API.  The API requires that at least one filter be given. See
'Details'}

\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{language}{\code{character(1)}. The language for which to look up 
domains.}

\item{url_base}{\code{character(1)}. The url stem for the API.}
}
\value{
A data frame with the following columns:

\itemize{
 \item \code{id } Word ID
 \item \code{word } The plain text of the word.
}
}
\description{
Get a list of words for a particular domain, lexical 
  category, register, and/or region.
}
\details{
The response only includes headwords, not all their possible 
inflections.

Filters must be given as named arguments in \code{...}. Each argument is a 
character vector. Examples include \code{registers = "Rare"} or 
\code{domains = c("Art", "Science")}.
}
\section{Functional Requirements}{

\enumerate{
 \item Return a data frame with the results of the API call.
 \item Cast an error if the API returns an error.
 \item Cast an error if any element in \code{...} is unnamed.
 \item Cast an error if any element in \code{...} is not a character vector.
 \item Cast an error if \code{app_id} is not \code{character(1)}.
 \item Cast an error if \code{app_id} is \code{NULL}.
 \item Cast an error if \code{app_key} is not \code{character(1)}.
 \item Cast an error if \code{app_key} is \code{NULL}.
 \item Cast an error if \code{language} is not \code{character(1)}.
 \item Cast an error if \code{url_base} is not \code{character(1)}.
 \item Cast an error if any element in \code{advanced_filter} is unnamed.
 \item Cast an error if any element in \code{advanced_filter} is not a character vector.
}
}

