% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxford_error.R
\name{oxford_error}
\alias{oxford_error}
\title{Process Errors from the Oxford Dictionaries API}
\usage{
oxford_error(response)
}
\arguments{
\item{response}{An object of class \code{response} as generated by a 
\code{\link[httr]{GET}} call.}
}
\description{
Provides a consistent interface for managing errors from the 
  Oxford Dictionaries API
}
