% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_application_access.R
\name{set_application_access}
\alias{set_application_access}
\alias{set_oxford_url_base}
\alias{set_application_id}
\alias{set_application_key}
\alias{set_application_language}
\alias{get_application_id}
\alias{get_application_key}
\alias{get_oxford_url_base}
\alias{get_application_language}
\title{Set Application Access Parameters}
\usage{
set_application_access(app_id, app_key)

set_oxford_url_base(url_base)

set_application_id(app_id)

set_application_key(app_key)

set_application_language(language = getOption("oxford_api_language"))

get_application_id()

get_application_key()

get_oxford_url_base()

get_application_language()
}
\arguments{
\item{app_id}{\code{character(1)}. The user's application ID.}

\item{app_key}{\code{character(1)}. The user's application key.}

\item{url_base}{\code{character(1)}. Sets the base of the URL to the 
Oxford Dictionaries API.}
}
\description{
Access to the Oxford Dictionaries API requires that 1) the 
  URL to the API is properly constructed, 2) the call utilizes a valid
  application ID, and 3) the call utilizes a valid application key.
}
\section{Functional Requirements}{


\code{set_application_access}

\enumerate{
 \item Successfully sets the value of \code{options("oxford_api_app_id")}
 \item Successfully sets the value of \code{options("oxford_api_app_key")}
 \item Casts an error if \code{app_id} is not a \code{character(1)}
 \item Casts an error if \code{app_id} does not have exactly eight characters.
 \item Casts an error if \code{app_key} is not a \code{character(1)}
 \item Casts an error if \code{app_key} does not have exactly 32 characters.
}

\code{set_oxford_url_base}

\enumerate{
 \item Successfully sets the value of \code{options("oxford_api_url_base")}
 \item Casts an error if \code{url_base} is not a \code{character(1)}
}

\code{set_application_id}

\enumerate{
 \item Successfully sets the value of \code{options("oxford_api_app_id")}
 \item Casts an error if \code{app_id} is not a \code{character(1)}
 \item Casts an error if \code{app_id} does not have exactly eight characters.
}

\code{set_application_key}

\enumerate{
 \item Successfully sets the value of \code{options("oxford_api_app_key")}
 \item Casts an error if \code{app_key} is not a \code{character(1)}
 \item Casts an error if \code{app_key} does not have exactly 32 characters.
}

\code{set_application_language}

\enumerate{
 \item Successfully sets the value of \code{options("oxford_api_language")}
 \item Casts an error if \code{language} is not one of 
   \code{oxford_languages} (See `?oxford_languages`)
}

\code{get_oxford_url_base}

\enumerate{
 \item Successfully retrieves the value of \code{options("oxford_api_url_base")}
}

\code{get_application_id}

\enumerate{
 \item Successfully retrieves the value of \code{options("oxford_api_app_id")}
}

\code{get_application_key}

\enumerate{
 \item Successfully retrieves the value of \code{options("oxford_api_app_key")}
}

\enumerate{
 \item Successfully retrieves the value of \code{options("oxford_api_language")}
}
}

\examples{

set_application_access(app_id = "myAppId1",
                       app_key = "This_is_not_a_real_key_123456789")

}
