% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_grammatical_feature.R
\name{tidy_grammatical_features}
\alias{tidy_grammatical_features}
\title{Organize Grammatical Features into a Data Frame}
\usage{
tidy_grammatical_features(gf)
}
\arguments{
\item{gf}{a list of grammatical features.}
}
\value{
A data frame with columns \code{grammatical_feature} and 
\code{grammatical_feature_type}.
}
\description{
Lexical entries may have multiple grammatical features.
  This function organizes these into a data frame with one row per 
  features.  This is a necessary step toward organizing a lexical entry.
}
