% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_inflection_of.R
\name{tidy_inflection_of}
\alias{tidy_inflection_of}
\title{Organize Word Inflections into a Data Frame}
\usage{
tidy_inflection_of(inflection)
}
\arguments{
\item{inflection}{A list containing inflections of a word.}
}
\description{
A word may have multiple inflections.  This function 
  organizes them into a data frame that may be joined with other 
  lexical entry information.
}
