% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lexical_entry.R
\name{tidy_lexical_entry}
\alias{tidy_lexical_entry}
\title{Organize a Lexical Entry into a Data Frame}
\usage{
tidy_lexical_entry(le)
}
\arguments{
\item{le}{A list containing a lexical entry}
}
\description{
A word may have multiple lexical entries. This function 
  gathers information about the entries into a data frame that is 
  suitable for indexing, filtering, etc.
}
